#123 Change directory of Repository and DataSource
