+++ app/src/main/java/com/webrtc/boyj/api/firebase/MyFirebaseMessagingService.java
@@ -7,7 +7,7 @@ import android.support.annotation.NonNull;
 
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
-import com.webrtc.boyj.data.source.preferences.TokenDataSource;
+import com.webrtc.boyj.data.source.TokenDataSource;
 import com.webrtc.boyj.data.source.preferences.TokenDataSourceImpl;
 import com.webrtc.boyj.presentation.ringing.RingingActivity;
