+++ app/src/main/java/com/webrtc/boyj/data/source/TokenDataSource.java
@@ -1,0 +1,14 @@
+package com.webrtc.boyj.data.source;
+
+import android.support.annotation.NonNull;
+
+public interface TokenDataSource {
+
+    boolean isNewToken();
+
+    void registerToken(@NonNull final String token);
+
+    String getToken();
+
+    void unsetNewToken();
+}
