+++ app/src/main/java/com/webrtc/boyj/data/source/remote/response/UserResponse.java
@@ -1,0 +1,4 @@
+package com.webrtc.boyj.data.source.remote.response;
+
+public class UserResponse {
+}
